added confirm on delete

+++ scripts/DailyJournalComponents/SavedJournalEntryListComponent.js
@@ -1,8 +1,4 @@
-import {
-  useSavedJournalEntries,
-  getEntries,
-  deleteEntries
-} from "./JournalDataProviderComponent.js";
+import { useSavedJournalEntries, getEntries, deleteEntries } from "./JournalDataProviderComponent.js";
 import SavedJournalEntryComponent from "./SavedJournalEntryComponent.js";
 
 const SavedJournalEntryListComponent = () => {
@@ -25,11 +21,13 @@ const SavedJournalEntryListComponent = () => {
 
   eventHub.addEventListener("click", clickEvent => {
     if (clickEvent.target.id.startsWith("deleteEntry--")) {
-      const [prefix, id] = clickEvent.target.id.split("--");
-      deleteEntries(id).then(() => {
-        const newEntry = useSavedJournalEntries();
-        renderData(newEntry);
-      });
+      if (confirm("Are you sre you want to delete this?")) {
+        const [prefix, id] = clickEvent.target.id.split("--");
+        deleteEntries(id).then(() => {
+          const newEntry = useSavedJournalEntries();
+          renderData(newEntry);
+        });
+      }
     }
   });
 
@@ -99,9 +97,7 @@ const SavedJournalEntryListComponent = () => {
     entryHTML.innerHTML = `
         <section>
           <h2 class="notes-title">Entries:</h2>
-          ${entriesCollection
-            .map(entry => SavedJournalEntryComponent(entry))
-            .join("")}
+          ${entriesCollection.map(entry => SavedJournalEntryComponent(entry)).join("")}
         </section>
       `;
   };
